change translation processing for stubborn lupdate
The lupdate tool is somewhat stubborn when processing project files with
subdirs: It doesn't work if there are no TRANSLATIONS named in the main
project file, and it is confused if there are TRANSLATIONS named in the
subdirs (even if they are the same). But we need to know TRANSLATIONS in
the subdirs to automatically build *.qm of them.

Solution is to avoid the TRANSLATIONS keyword in subdir project files
and just to use another variable name. What the heck ...